Moved PatchRequest.Operation out as a class of its own. Updated to Jackson 2.5.0.
