+++ pom.xml
@@ -29,7 +29,7 @@
         <compileSource>1.6</compileSource>
         <main.basedir>${project.basedir}</main.basedir>
         <ignore.test.failures>false</ignore.test.failures>
-        <com.fasterxml.jackson.version>2.0.0</com.fasterxml.jackson.version>
+        <com.fasterxml.jackson.version>2.5.0</com.fasterxml.jackson.version>
     </properties>
 
     <distributionManagement>
